misc: add wgsl target and move module printing test case to filecheck
