+++ xdsl/interpreters/wgpu.py
@@ -5,12 +5,12 @@ from typing import Any, cast
 import wgpu  # pyright: ignore[reportMissingTypeStubs]
 import wgpu.utils  # pyright: ignore[reportMissingTypeStubs]
 
+from xdsl.backend.wgsl.wgsl_printer import WGSLPrinter
 from xdsl.dialects import gpu
 from xdsl.dialects.builtin import IndexType
 from xdsl.dialects.memref import MemRefType
 from xdsl.interpreter import Interpreter, InterpreterFunctions, impl, register_impls
 from xdsl.interpreters.shaped_array import ShapedArray
-from xdsl.interpreters.wgsl_printer import WGSLPrinter
 from xdsl.ir import Attribute, SSAValue
 from xdsl.traits import SymbolTable
 from xdsl.utils.hints import isa
